+++ CHANGELOG.rst
@@ -12,7 +12,7 @@ Major Features and Improvements
 Breaking changes
 ------------------
 
-Depreceations
+Deprecations
 -------------
 
 
@@ -21,6 +21,7 @@ Bug fixes and small changes
 - fix color wrong in printout of results, params
 - packaging: moved to pyproject.toml and a setup.cfg mainly, development requirements can
   be installed with the `dev` extra as (e.g.) `pip install zfit[dev]`
+- add correlation method to FitResult
 
 Experimental
 ------------
@@ -124,7 +125,7 @@ Breaking changes
 - Integrals of extended PDFs are not extended anymore, but `ext_integrate` now returns the
   integral multiplied by the yield.
 
-Depreceations
+Deprecations
 -------------
 - `ComposedParameter` takes now `params` instead of `dependents` as argument, it acts now as
   the arguments to the `value_fn`. To stay future compatible, create e.g. `def value_fn(p1, pa2)`